Java Train Graph
Added in a test to check if the children are present. The test however does not mention the api in any way. I'm really trying to be stricter on behavior here. If I can get that working correctly then I don't have to worry about what code I write as long as I pass the test.