Java Train Graph
TDD refactor checkpoint: refactored the code to use a traditional loop instead of a loop that initalizes the node on each run. I hope this will help runtime performance for this "toy" code. Also added in a break once a node is found. This will remove the first child in the list of children.

[Question] Interestingly enough if two nodes share the same name which one will be deleted? For now I'm not going to worry about that design decision as I'm going to throw this code away, but it is something to think about.