Java Train Graph
Oops... forgot to commit after a successful test pass. But this is a plus. I'm following TDD. I wrote a test that passed and did not refactor or change the code at all. This allowed me to look at my assertions and not have to worry about anything else. The case I supplied my code passed so I can move on to other things. This will probably be caught in mutation testing if I were to run it, but that is still ok as we can write the test to handle it then.

[improve] Need to keep up commiting once I finish a cycle. Two passing tests in a row does not forgive this behvior.