Java Train Graph
Added the ability to add a vertex to my list. This is all well and good, but I think its important to also look at what's happening here. I have to know the location in my matrix to on where to put my vertex. This means that I am potentially exposing my state to the world.

What can I do to make sure that I'm encapsulating state in this example and making sure that I can do what I need to do. I don't like the print() method, but for now it is good enough to visually see what is happening to make sure the tests pass. I will need to refactor the tests to reflect the real behavior I am trying to caputure.