Java Train Graph
Added in a failsafe for adding a vetex to the graph. This way I don't have to do any magic when handling an insert. Since we are a square graph (for now) we don't have to worry as much aobut adding something to the list. We can check along the axis. If we know what we are adding we can do what we need to.